LC-powx-n [Runtime: N/A; Memory: N/A]

+++ math/powx-n.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     double myPow(double x, int n) {
         double ans = 1.0;
-        int num = n;
+        long long num = n;
         if(num < 0) num = num * -1;
         while(num) {
             if(num % 2 == 1) {
